+++ airbase/__init__.py
@@ -1,7 +1,7 @@
 from . import resources, util
 from .__version__ import __version__
 from .airbase import AirbaseClient, AirbaseRequest
-from .download_api import Dataset
+from .parquet_api import Dataset
 
 __all__ = [
     "AirbaseClient",
